requirements.txt
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
environs==11.0.0
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
marshmallow==3.21.3
multidict==6.0.5
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
